+++ test/types/schema.test.ts
@@ -21,11 +21,9 @@ import {
   Types,
   Query,
   model,
-  ValidateOpts,
-  BufferToBinary
+  ValidateOpts
 } from 'mongoose';
-import { Binary } from 'mongodb';
-import { IsPathRequired } from '../../types/inferschematype';
+import { Binary, BSON } from 'mongodb';
 import { expectType, expectError, expectAssignable } from 'tsd';
 import { ObtainDocumentPathType, ResolvePathType } from '../../types/inferschematype';
 
@@ -590,6 +588,16 @@ const batchSchema2 = new Schema({ name: String }, { discriminatorKey: 'kind', st
   return 1;
 } } });
 batchSchema2.discriminator('event', eventSchema2);
+
+
+function encryptionType() {
+  const keyId = new BSON.UUID();
+  expectError<Schema>(new Schema({ name: { type: String, encrypt: { keyId } } }, { encryptionType: 'newFakeEncryptionType' }));
+  expectError<Schema>(new Schema({ name: { type: String, encrypt: { keyId } } }, { encryptionType: 1 }));
+
+  expectType<Schema>(new Schema({ name: { type: String, encrypt: { keyId } } }, { encryptionType: 'queryableEncryption' }));
+  expectType<Schema>(new Schema({ name: { type: String, encrypt: { keyId } } }, { encryptionType: 'csfle' }));
+}
 
 function gh11828() {
   interface IUser {
@@ -1724,4 +1732,25 @@ async function gh12959() {
   expectType<number>(doc.__v);
   const leanDoc = await TestModel.findOne().lean().orFail();
   expectType<number>(leanDoc.__v);
+}
+
+async function gh15236() {
+  const schema = new Schema({
+    myNum: { type: Number }
+  });
+
+  schema.path<Schema.Types.Number>('myNum').min(0);
+}
+
+function gh15244() {
+  const schema = new Schema({});
+  schema.discriminator('Name', new Schema({}), { value: 'value' });
+}
+
+async function schemaDouble() {
+  const schema = new Schema({ balance: 'Double' });
+  const TestModel = model('Test', schema);
+
+  const doc = await TestModel.findOne().orFail();
+  expectType<Types.Double | null | undefined>(doc.balance);
 }